+++ test/unit/carriers/benchmark_test.rb
@@ -1,7 +1,7 @@
 require 'test_helper'
 
 class BenchmarkTest < ActiveSupport::TestCase
-  include ActiveShipping::Test::Fixtures
+  include ReactiveShipping::Test::Fixtures
 
   def setup
     @carrier = BenchmarkCarrier.new